Remove XML importer
Firefox's Quantum update made many extensions, including the password exporter, irrevocably incompatible. Though some users may be on ESR or just haven't updated Firefox yet, given that the extension cannot even be obtained anymore (it has been unlisted/removed), there's no point to keeping the XML importer around and putting in the effort to make it compatible with the library refactor.

The XML import feature was redundant to begin with given that the other importer provides the same functionality but without a middleman.